Add assignment operations & add room options max assignment group count and operations enabled or not.

+++ API/Kollity.Application/Commands/Assignment/Group/DeleteStudent/DeleteStudentFromAssignmentGroupCommandHandler.cs
@@ -9,7 +9,7 @@ public class DeleteStudentFromAssignmentGroupCommandHandler : ICommandHandler<De
     private readonly ApplicationDbContext _context;
     private readonly IUserAccessor _userAccessor;
 
-    public DeleteStudentFromAssignmentGroupCommandHandler(ApplicationDbContext context,IUserAccessor userAccessor)
+    public DeleteStudentFromAssignmentGroupCommandHandler(ApplicationDbContext context, IUserAccessor userAccessor)
     {
         _context = context;
         _userAccessor = userAccessor;
@@ -21,14 +21,23 @@ public class DeleteStudentFromAssignmentGroupCommandHandler : ICommandHandler<De
         Guid userId = _userAccessor.GetCurrentUserId(),
             groupId = request.InvitationDto.GroupId,
             studentId = request.InvitationDto.StudentId;
-        
+
+        var roomOperationsState = await _context.Rooms
+            .Where(x => x.AssignmentGroups.Any(xx => xx.Id == groupId))
+            .Select(x => x.AssignmentGroupOperationsEnabled)
+            .FirstOrDefaultAsync(cancellationToken);
+
+        if (roomOperationsState == false)
+            return AssignmentErrors.OperationIsOff;
+
+
         //check user is in the group
         var isJoined = await _context.AssignmentGroupStudents
             .AnyAsync(x => x.StudentId == userId && x.AssignmentGroupId == groupId && x.JoinRequestAccepted,
                 cancellationToken);
         if (isJoined == false)
             return AssignmentErrors.UserIsNotInTheGroup;
-        
+
         await _context.AssignmentGroupStudents
             .Where(x => x.StudentId == studentId && x.AssignmentGroupId == groupId && x.JoinRequestAccepted)
             .ExecuteDeleteAsync(cancellationToken);
